cleanup

+++ Content.Shared/Conditions/UnifiedConditions/IAntagonistTagCondition.cs
@@ -1,21 +1,20 @@
 using System.Linq;
 using Content.Shared.Conditions.HelperConditions;
-using Content.Shared.EntityConditions.Conditions.Mind;
 using Content.Shared.Mind;
 using Content.Shared.Roles;
 using Robust.Shared.Prototypes;
 
 namespace Content.Shared.Conditions.UnifiedConditions;
+
 /// <summary>
 /// Checks if the given mind is an antagonist with specified tag.
 /// </summary>
 public interface IAntagonistTagCondition : ICondition<IAntagonistTagCondition>, IWithInverted
 {
-
     /// <summary>
     /// The tags this check will succeed for.
     /// For example, if "OnStation" is provided, all on-station antags will pass the check.
-    /// If <see cref="AllowNonAntags"/> is true, it will additionally allow every non-antag to pass the check.
+    /// If <see cref="AllowNonAntags" /> is true, it will additionally allow every non-antag to pass the check.
     /// </summary>
     HashSet<ProtoId<AntagTagPrototype>> Tags { get; }
 
@@ -24,7 +23,6 @@ public interface IAntagonistTagCondition : ICondition<IAntagonistTagCondition>, 
     /// </summary>
 
     bool AllowNonAntags { get; }
-
 }
 
 public sealed partial class AntagonistTagEntityConditionSystem : EntitySystem
@@ -43,6 +41,6 @@ public sealed partial class AntagonistTagEntityConditionSystem : EntitySystem
             return;
         }
 
-        args.Value = (float)antagTags.Intersect(conditionTags).Count() / (float)args.Condition.Tags.Count;
+        args.Value = antagTags.Intersect(conditionTags).Count() / (float)args.Condition.Tags.Count;
     }
 }
