+++ Content.Shared/Conditions/UnifiedConditions/IJobCondition.cs
@@ -16,7 +16,8 @@ public interface IJobCondition : ICondition<IJobCondition>
 }
 
 /// <summary>
-/// Returns true if this entity has any of the specified jobs. False if the entity has no mind, none of the specified jobs, or is jobless.
+/// Returns true if this entity has any of the specified jobs. False if the entity has no mind, none of the specified jobs,
+/// or is jobless.
 /// </summary>
 public sealed partial class JobConditionSystem : EntitySystem
 {
@@ -27,7 +28,7 @@ public sealed partial class JobConditionSystem : EntitySystem
     {
         args.Handled = true;
 
-        args.Value = args.Condition.Jobs.Count(job=>_job.MindHasJobWithId(entity.Comp.Mind,job));
+        args.Value = args.Condition.Jobs.Count(job => _job.MindHasJobWithId(entity.Comp.Mind, job));
     }
 
     [SubscribeLocalEvent]
@@ -35,6 +36,7 @@ public sealed partial class JobConditionSystem : EntitySystem
     {
         args.Handled = true;
 
-        args.Value = (float)args.Condition.Jobs.Count(job=>_job.MindHasJobWithId(entity,job)) / (float)args.Condition.Jobs.Length;
+        args.Value = args.Condition.Jobs.Count(job => _job.MindHasJobWithId(entity, job)) /
+                     (float)args.Condition.Jobs.Length;
     }
 }
